block key value size use uint32_t instead of size_t

+++ src/Map.h
@@ -5,6 +5,8 @@
 #ifndef SRC_MAP_H_
 #define SRC_MAP_H_
 
+#include <cstddef>
+#include <cstdint>
 #include <ctime>
 #include <string>
 #include <unordered_map>
@@ -14,7 +16,7 @@ namespace tinydb {
 struct Value {
   std::uint64_t file_id;
   std::size_t value_size;
-  std::size_t offset;
+  std::uint64_t offset;
   std::time_t timestamp;
 };
